Last commit

+++ philo/src/main.c
@@ -6,7 +6,7 @@
 /*   By: sguilher <sguilher@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/15 23:57:12 by sguilher          #+#    #+#             */
-/*   Updated: 2022/08/12 00:10:51 by sguilher         ###   ########.fr       */
+/*   Updated: 2022/08/12 01:53:27 by sguilher         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,18 +14,20 @@
 
 void	simulation_monitoring(t_args *philos, t_data *data)
 {
-	int	i;
-	int	delta;
+	int			i;
+	long long	delta;
 
 	while (simulation(data) != STOP)
 	{
 		i = 0;
+		//printf("last_eat = %lli\n", philos[i].last_eat);
 		while (i < data->nbr_of_philos)
 		{
 			if (simulation(data) == STOP)
 				break ;
 			pthread_mutex_lock(&(philos[i].lock_philo));
 			delta = time_now() - philos[i].last_eat;
+			//printf("delta = %i\n", delta);
 			if (delta >= data->time.to_die)
 			{
 				pthread_mutex_lock(&(data->lock_data));
@@ -67,8 +69,3 @@ int	main(int argc, char *argv[])
 	// limpar forks, philos, args
 	return (0);
 }
-
-// casos para testar:
-// tempos = 0
-// 5 100 0
-// 5 0 0
